requirements.txt
aiocontextvars==0.2.2
anyio==3.6.2
asgiref==3.4.1
asn1crypto==0.24.0
click==8.0.4
contextlib2==21.6.0
contextvars==2.4
cryptography==2.1.4
dataclasses==0.8
Deprecated==1.2.13
fastapi==0.83.0
fastapi-utils==0.2.1
greenlet==2.0.1
h11==0.13.0
httptools==0.3.0
idna==3.4
immutables==0.19
importlib-metadata==4.8.3
keyring==10.6.0
keyrings.alt==3.0
opentelemetry-api==1.12.0
opentelemetry-instrumentation==0.33b0
opentelemetry-instrumentation-asgi==0.33b0
opentelemetry-instrumentation-fastapi==0.33b0
opentelemetry-semantic-conventions==0.33b0
opentelemetry-util-http==0.33b0
pycrypto==2.6.1
pydantic==1.9.2
PyGObject==3.26.1
python-dotenv==0.20.0
pyxdg==0.25
PyYAML==6.0
SecretStorage==2.3.1
six==1.11.0
sniffio==1.2.0
SQLAlchemy==1.4.44
starlette==0.19.1
typing_extensions==4.1.1
uvicorn==0.16.0
uvloop==0.14.0
watchgod==0.7
websockets==9.1
wrapt==1.14.1
zipp==3.6.0